+++ lib/DietMonitoring/diet_monitoring_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 // Import the location of your Survey models
 import 'package:namer_app/Survey/survey_page.dart'; // Import your SurveyPage widget
 import 'package:namer_app/models/task_models.dart';
-import 'package:namer_app/Tasks/Survey/tasks.dart'; // Import tasks related to surveys
+import 'package:namer_app/constants/tasks/Survey/tasks.dart'; // Import tasks related to surveys
 
 // Assuming 'themeColor' is defined elsewhere in your application
 const Color themeColor = Colors.blue; // Placeholder for the actual theme color
