+++ lib/pages/Login/register_info_page.dart
